Move RegistrationView from user_api to user_authn.
